Fixed: When several levels of subfolders are created rapidly, the watcher gets lost and doesn't follows the tree up to the leaves
