+++ src/main/scala/bss/EventsRouter.scala
@@ -3,7 +3,7 @@ package bss
 import java.nio.file.Path
 
 import akka.actor.{ Props, ActorLogging, Actor }
-import bss.WatcherEvents.PathEvent
+import bss.RecursiveWatcher.PathEvent
 
 object EventsRouter {
   def props(rootPath: Path) = Props(classOf[EventsRouter], rootPath)
@@ -14,6 +14,6 @@ class EventsRouter(rootPath: Path) extends Actor with ActorLogging {
   def receive = {
     case PathEvent(eventType, path, isDirectory, count) =>
       val relPath = rootPath.relativize(path)
-      log.info(s">>> $eventType ($count) : $relPath")
+      log.info(s">>> $eventType : $relPath")
   }
 }
